refactor: move k3s cluster nginx code to a dedicated file
For better modularity and readability.
