+++ core/src/plugins/kubernetes/nginx/ingress-controller.ts
@@ -14,9 +14,9 @@ import { microk8sNginxInstall, microk8sNginxStatus, microk8sNginxUninstall } fro
 import { apply } from "../kubectl.js"
 import { minikubeNginxInstall, minikubeNginxStatus, minikubeNginxUninstall } from "../local/minikube.js"
 import { defaultBackendStatus } from "./default-backend.js"
-import { getK3sNginxHelmValues } from "../local/k3s.js"
 import { getGenericNginxHelmValues } from "./nginx-helm-generic.js"
 import { helmNginxInstall, helmNginxStatus, helmNginxUninstall } from "./nginx-helm.js"
+import { getK3sNginxHelmValues } from "./nginx-helm-k3s.js"
 
 export async function ingressControllerReady(ctx: KubernetesPluginContext, log: Log): Promise<boolean> {
   const clusterType = ctx.provider.config.clusterType
